POPPY-0080 Common, payload-related functions are moved to SRPayloads class
