+++ src/core/apps/ServiceRegistry/payloads/SRSystem.h
@@ -3,13 +3,12 @@
 
 #include <string>
 #include <vector>
-#include "commonPayloads.h"
+#include "SRPayloads.h"
 #include "gason/gason.h"
 #include "../utils/SRJsonBuilder.h"
-
 #include <ctype.h>
 
-class SRSystem
+class SRSystem : SRPayloads
 {
     private:
 
@@ -26,22 +25,6 @@ class SRSystem
         std::string sAuthInfo;
         std::string sCreatedAt;
         std::string sUpdatedAt;
-
-        void toLowerAndTrim(std::string &_s)
-        {
-            std::transform(_s.begin(), _s.end(), _s.begin(), ::tolower);
-
-            const char *whitespace = " \n\r\t\f\v";
-            const char *s2 = _s.c_str();
-            int n = _s.size();
-            char *tmp = (char *)malloc(n+1);
-            int j = 0;
-            for(int i = 0; i < n; ++i)
-                if( strchr(whitespace, s2[i]) == NULL) tmp[j++] = s2[i];
-            tmp[j] = '\0';
-            _s = std::string(tmp);
-            free(tmp);
-        }
 
         bool setJsonPayload(std::string &_sJsonPayload)
         {
